+++ R/brace_linter.R
@@ -8,9 +8,12 @@
 #'  - Closing curly braces in `if` conditions are on the same line as the corresponding `else`.
 #'  - Either both or neither branch in `if`/`else` use curly braces, i.e., either both branches use `{...}` or neither
 #'    does.
-#'  - Functions spanning multiple lines use curly braces.
+#'  - Function bodies are wrapped in curly braces.
 #'
-#' @param allow_single_line if `TRUE`, allow an open and closed curly pair on the same line.
+#' @param allow_single_line If `TRUE`, allow an open and closed curly pair on the same line.
+#' @param function_braces Character scalar specifying whether to require function bodies to be wrapped in curly braces.
+#'   `"multi_line"` means to only require curly braces when a function body is not on the same line as its header or it
+#'   spans over multiple lines.
 #'
 #' @examples
 #' # will produce lints
@@ -50,7 +53,10 @@
 #' - <https://style.tidyverse.org/syntax.html#indenting>
 #' - <https://style.tidyverse.org/syntax.html#if-statements>
 #' @export
-brace_linter <- function(allow_single_line = FALSE) {
+brace_linter <- function(allow_single_line = FALSE,
+                         function_braces = c("multi_line", "always", "never")) {
+  function_braces <- match.arg(function_braces)
+
   xp_cond_open <- xp_and(c(
     # matching } is on same line
     if (isTRUE(allow_single_line)) {
@@ -124,7 +130,9 @@ brace_linter <- function(allow_single_line = FALSE) {
   # TODO (AshesITR): if c_style_braces is TRUE, this needs to be @line2 + 1
   xp_else_same_line <- glue("//ELSE[{xp_else_closed_curly} and @line1 != {xp_else_closed_curly}/@line2]")
 
-  xp_function_brace <- "(//FUNCTION | //OP-LAMBDA)/parent::expr[@line1 != @line2 and not(expr[OP-LEFT-BRACE])]"
+  xp_function_brace_always <- "(//FUNCTION | //OP-LAMBDA)/parent::expr[not(expr[OP-LEFT-BRACE])]"
+  xp_function_brace_multi_line <-
+    "(//FUNCTION | //OP-LAMBDA)/parent::expr[@line1 != @line2 and not(expr[OP-LEFT-BRACE])]"
 
   # if (x) { ... } else if (y) { ... } else { ... } is OK; fully exact pairing
   #   of if/else would require this to be
@@ -192,14 +200,25 @@ brace_linter <- function(allow_single_line = FALSE) {
       )
     )
 
-    lints <- c(
-      lints,
-      xml_nodes_to_lints(
-        xml_find_all(xml, xp_function_brace),
-        source_expression = source_expression,
-        lint_message = "Any function spanning multiple lines should use curly braces."
+    if (function_braces == "always") {
+      lints <- c(
+        lints,
+        xml_nodes_to_lints(
+          xml_find_all(xml, xp_function_brace_always),
+          source_expression = source_expression,
+          lint_message = "Any function body should be wrapped in curly braces."
+        )
       )
-    )
+    } else if (function_braces == "multi_line") {
+      lints <- c(
+        lints,
+        xml_nodes_to_lints(
+          xml_find_all(xml, xp_function_brace_multi_line),
+          source_expression = source_expression,
+          lint_message = "Any function body spanning multiple lines should be wrapped in curly braces."
+        )
+      )
+    }
 
     lints <- c(
       lints,